Move implementation to logs, and add (deprecated) aliases.
